GitCaller fully working
Fixed some pieces of code that were using the jGit library incorrectly. After
resolving those problems, I tried to use the RevFilter class to get commits with
timestamps in-between the since and until values in
GitCaller.getPartialCommits(), but it wasn't behaving as I expected and I
couldn't figure out why not. As a result, I added the
GitCaller.isCommitInDateRange() function as a work-around. To be honest, I think
(althogh I'm not sure) that isCommitInDateRange() does almost exactly what jGit
would have done internally had I used a RevFilter, which is the more typical
solution, so I don't think the workaround even adds a performance penalty. We
might never have any reason to revisit this implementation and have it use a
RevFilter instead.