Refactor design so future tasks are easier
In the future, we want to turn our program into an Eclipse plugin that will let
users know about imminent and past conflicts. To do so, we think it would be
best to load all of the target repository's commits in one startup step, then
repeatedly use that data to alow the user to sonify different date ranges.

In order to make sure that future task will be easier to accomplish, this commit
refactors the program so that the steps of getting raw data from the target
repository and selecting which commits will actually be used for sonification
are decoupled. Previously, it would have been difficult to modify the program
so that if you wanted to sonify multiple date ranges, the target repository's
.git directory would only have to be opened once; now, it is much easier to
accomplish that.

This commit does not change the program's visible behavior.