Using Joda Time (http://www.joda.org/joda-time/index.html)
For one of the steps in preparing Commits to be sonified, we need to calculate
the number of days between two dates. This looks painfully tricky to do with
Java's built-in Date and Calendar classes, so we're now using the Joda Time
library instead. This commit adds the library and refactors our existing code so
that everywhere that previously used java.util.Date now uses Joda Time classes
instead.