Implemented MeasureProducer for per commit sonification
Implemented the logic that produces measures that can be used for the per
commit sonification.

I realized that the two different sonifications have different needs for the
measures they consume, so the MeasureProducer class is now abstract. In
addition, I changed the name of the MeasureProduce class (it was previously
CommitProcessor) because there's another class in our program called
CommitProcessor, and I want to avoid confusion in the future.